scripts/requirements.txt
# FFF Thermal Simulation Requirements

# Base scientific computing
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.3.0
pandas>=1.3.0
pyyaml>=5.4.0

# GPU Support (choose one based on your CUDA version)
# For CUDA 11.x:
# cupy-cuda11x

# For CUDA 12.x:
# cupy-cuda12x

# Alternative GPU support via PyTorch:
# torch
# torchvision  
# torchaudio
# --index-url https://download.pytorch.org/whl/cu118

# Development and testing
pytest>=6.0.0
jupyter>=1.0.0

# Optional: Enhanced visualization
plotly>=5.0.0
seaborn>=0.11.0
